1 - Fixed prefix crit boosts not working
2 - Swapped effects of Intrepid and Hasty
3 - Removed two steps of Dauntless HP scaling

+++ Prefixes/VanillaPrefixChanges/Stats/PrefixCritChanceStat.cs
@@ -3,9 +3,9 @@ using Terraria.ModLoader;
 
 namespace CalamityMod.Prefixes.VanillaPrefixChanges.Stats
 {
-    public struct PrefixCritChanceStat(float critChanceBonus) : IVanillaPrefixStat
+    public struct PrefixCritChanceStat(int critChanceBonus) : IVanillaPrefixStat
     {
-        public float CritChanceBonus = critChanceBonus;
+        public int CritChanceBonus = critChanceBonus;
 
         public readonly void ApplyEffects(Player player)
         {
